[Izza] My User Fav - Change Database Dao for get and delete User

+++ app/src/main/java/id/co/iconpln/controlflowapp/database/FavoriteDatabaseDao.kt
@@ -12,12 +12,12 @@ interface FavoriteDatabaseDao {
     fun insertUser(user: FavoriteUser)
 
     @Query("DELETE FROM fav_user_table WHERE userId = :key")
-    fun deleteUser(key: Long)
+    fun deleteUser(key: Int)
 
     @Query("SELECT * FROM fav_user_table ORDER BY favUserId DESC")
     fun getAllUsers(): LiveData<List<FavoriteUser>>
 
     //get a specific user based on its id
     @Query("SELECT * from fav_user_table WHERE userId = :key")
-    fun getFavUser(key: Long): LiveData<FavoriteUser>
+    fun getFavUser(key: Int): LiveData<FavoriteUser>
 }
